better base handling, especially binary
Base 2 was basically being ignored. It just printed numbers
in base 10, which is really different than base 2. This allows
numbers to be entered with a specific base, regardless of what
the default base is. That includes binary, which did not have
a parser for 0b10101 style numbers before or a way to print
them out.

All this is fixed and more!

Signed-off-by: Vernon Mauery <vernon@mauery.org>