fixups to get all numeric backends working again
Each one has little issues that needed to be addressed. I had been
focusing work on clang/mpfr because that is the best choice. But I think
native types had not worked for some time and while they are not as
interesting because of the numeric limits, it is the most portable.

This also adds a script that will kick off a build with each of the
CC/numeric options to make sure it is all good.

Signed-off-by: Vernon Mauery <vernon@mauery.org>