Auto downcast of items on the stack
As items get pushed into stack entries, the code attempts
to losslessly downcast types. The hardest one is to go
from mpf to mpq. The rule is that it must be done at at
least 10^(-precision) error with precision/5 as the max
denominator. This strikes a good balance of exactness with
actually reducing known fractions.

 * mpc -> mpf for imaginary = 0
 * mpf -> mpz if no fractional part
 * mpf -> mpq for reasonable error
 * mpq -> mpz for denominator = 1

Signed-off-by: Vernon Mauery <vernon@mauery.org>