Add time types (duration and absolute)
Add time numeric types. Time point or absolute and duration. This
implementation is kind of a hack, but the c++20 stuff is not available.
This gets the time numeric type in there, which is nice.

This adds 'now' as a function that returns the current timestamp.
This also adds duration time literal parsers for times like:
 * 1s - 1 second
 * 3.14e6d - approximately pi million days
 * ns us ms s m h d

This means you can run:
    > now 90d +
to have it calculate what day is 90 days from now.

Signed-off-by: Vernon Mauery <vernon@mauery.org>