Add basic unit support
Use units by adding a _<unit> tag at the end of a number. The calculator
will attempt to enforce units stoichiometry for functions that make
sense, mostly arithmetic functions. Many functions will simply strip off
the units or require the values to be unitless.

The units are achieved by using prime numbers as unit IDs and then using
a quotient to keep track of the units. To determine what the units are,
it will factor the composite ID a print out each of the units. If the
unit matches a known unit type, it will resolve to that.

For example:
 > 1_kg*m/s 1_s /
 1: 1_N
 > 1_s /
 1: 1_m*kg/s*s*s

As you can see, it doesn't know what a N/s is, so it just prints all the
base SI units. If we had multiplied by 1_m instead, it knows a N*m is
actually a J, so it would have printed 1_J.

Signed-off-by: Vernon Mauery <vernon@mauery.org>