Change order in build-all.sh
This builds the trouble points sooner.

If one of the boost backends work, they usually all do, but the native
types (which is mildly interesting to me for embedded purposes) often is
at odds with the boost types.

Also, change the way builds are ignored. Meson adds a .gitignore to all
its build dirs, so this script should do the same.

Signed-off-by: Vernon Mauery <vernon@mauery.org>