break out numeric.hpp
numeric.hpp was getting too big and unwieldy. This pulls the basic types out
of numeric and the management of the types into either
numeric_basic_types or numeric_boost_types and puts the common
formatters into numeric_format.

It required a bit of reworking to get it to compile both native and
boost types, but it should work for both again.

Signed-off-by: Vernon Mauery <vernon@mauery.org>