Rewrote parser with boost::parser
boost::parser, which is not yet official, is supposed to
be slated for inclusion in the boost project as the
successor of spirit. While some of it is lame (such
as the inability to create a class around the parser)
it is nice to have all the symantic actions that
use the parsed parts to create the actual classes.

This also involved a lot of rewriting of the numeric
reparsing since the old parser basically only tokenized
the input. To keep parsing separate from numerics,
there are 'parser_parts' that are created, capturing and
labeling each of the important strings needed to create
the numeric classes and functions.

Since the top level parse is now a 'program' this also
introduces a change in how the calculator executes,
with each line as basically an in-line program and the
capability to execute a program that is saved on the
stack. It does not have a full language yet, but if
statements are the initial foray into execution
control.

Big changes.

Signed-off-by: Vernon Mauery <vernon@mauery.org>